Move some files to content-core module (#2)
