+++ content-core/src/main/java/com/devgary/contentcore/model/Content.kt
@@ -1,4 +1,4 @@
-package com.devgary.contentview.content
+package com.devgary.contentcore.model
 
 data class Content(
     val url: String,
