+++ content-core/src/main/java/com/devgary/contentcore/model/ContentType.kt
@@ -1,4 +1,4 @@
-package com.devgary.contentview.content
+package com.devgary.contentcore.model
 
 enum class ContentType {
     IMAGE,
