+++ content-view/build.gradle
@@ -36,6 +36,8 @@ android {
 }
 
 dependencies {
+    implementation project(path: ':content-core')
+
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.android.material:material:1.6.1'
