+++ content-view/src/main/java/com/devgary/contentview/content/ContentHandler.kt
@@ -1,5 +1,7 @@
 package com.devgary.contentview.content
 
+import com.devgary.contentcore.model.Content
+
 interface ContentHandler {
     fun showContent(content: Content)
 }
