+++ demo/build.gradle
@@ -37,6 +37,7 @@ android {
 
 dependencies {
     implementation project(path: ':content-view')
+    implementation project(path: ':content-core')
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.4.2'
